+++ src/stickler/structured_object_evaluator/models/confusion_matrix_builder.py
@@ -143,7 +143,6 @@ class ConfusionMatrixBuilder:
                     "tn": int,
                     "fn": int,
                     "similarity_score": float,
-                    "all_fields_matched": bool,
                     "derived": {  # If add_derived_metrics=True
                         "cm_precision": float,
                         "cm_recall": float,